src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/create-content-element-expected.txt

CONSOLE ERROR: line 27: Uncaught TypeError: target.createShadowRoot is not a function
Hello,contentWorld.
This test compares a shadow-based render tree with one for a reference DOM tree.
